pyproject.toml
[tool.poetry]
name = "microshop"
version = "0.1.0"
description = ""
authors = ["janetuyy <118744357+janetuyy@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
pydentic = {extras = ["email"], version = "^0.0.1.dev3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
async-timeout = "^4.0.3"
alembic = "^1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
